orders and details

+++ database/migrations/2017_08_20_115208_create_flower_packets_table.php
@@ -15,7 +15,7 @@ class CreateFlowerPacketsTable extends Migration
         Schema::create('flower_packets', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title');
-            $table->decimal('price', 19, 3)->nullable();
+            $table->float('price', 10, 3)->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
